[ADD] Added currency_id field, so that now it is possible to get accounting reports in another currency other than the company currency.
[IMP] Given that the new field was added
When no currency is selected the company currency 
is the default, is a different currency 
all the values will be expressed into that
new currency.

This behavior is now only available for
the four (4) columns report,

Will it be necessary to take into account which
date rate, the actual or another date for that
exchange.